LayoutTests/platform/mac/accessibility/aria-liveregion-without-element-access.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script>
var successfullyParsed = false;
</script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div role="group" tabindex=0 id="liveregion" aria-live="polite" aria-relevant="additions">
<h3 id="innerlive">text</h3>
<div role="group">
   <div id="addregion"></div>
</div>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that adding an element to a live region will trigger a notification even though the AX element was not created previously.");

    var liveRegion = 0;
    function ariaCallback(notification) {
        if (notification == "AXLiveRegionChanged") {
           alert("Successfully received " + notification);
           liveRegion.removeNotificationListener();
           window.layoutTestController.notifyDone();
        }
    }

    if (window.accessibilityController) {
        window.layoutTestController.waitUntilDone();

        document.getElementById("liveregion").focus();
        liveRegion = window.accessibilityController.focusedElement;

        // Trigger the live region callback for a new element. Add region has not been created
        // as an element, but this notification should still be sent for liveregion.
        document.getElementById("addregion").appendChild(document.createTextNode("added"));

        var addedNotification = liveRegion.addNotificationListener(ariaCallback);
        shouldBe("addedNotification", "true");
    }

    successfullyParsed = true;
</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
