LayoutTests/platform/mac/accessibility/aria-liveregion-without-element-access-expected.txt

ALERT: Successfully received AXLiveRegionChanged
text

added
This tests that adding an element to a live region will trigger a notification even though the AX element was not created previously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
